+++ src/torch_tools/models/_simple_conv_2d.py
@@ -1,6 +1,7 @@
 """A simple two-dimensional convolutional neural network."""
 from typing import Optional, Dict, Any
 
+from torch import Tensor
 from torch.nn import Sequential, Flatten
 
 from torch_tools.models._encoder_2d import Encoder2d
@@ -9,6 +10,7 @@ from torch_tools.models._fc_net import FCNet
 from torch_tools.models._conv_net_2d import _forbidden_args_in_dn_kwargs
 from torch_tools.models._argument_processing import process_num_feats
 from torch_tools.models._argument_processing import process_2d_kernel_size
+from torch_tools.models._argument_processing import process_2d_block_style_arg
 
 # pylint: disable=too-many-arguments
 
@@ -40,6 +42,9 @@ class SimpleConvNet2d(Sequential):
     fc_net_kwargs : Dict[str, Any], optional
         Keyword arguments for ``torch_tools.models.fc_net.FCNet`` which serves
         as the classification/regression part of the model.
+    block_style : str, optional
+        Style of encoding blocks to use: choose from ``"double_conv"`` or
+        ``conv_res``.
 
     Examples
     --------
@@ -68,6 +73,7 @@ class SimpleConvNet2d(Sequential):
         lr_slope: float = 0.1,
         kernel_size: int = 3,
         fc_net_kwargs: Optional[Dict[str, Any]] = None,
+        block_style: str = "double_conv",
     ):
         """Build ``SimpleConvNet2d``."""
         encoder_feats = self._num_output_features(
@@ -88,6 +94,7 @@ class SimpleConvNet2d(Sequential):
                 downsample_pool,
                 lr_slope,
                 process_2d_kernel_size(kernel_size),
+                block_style=process_2d_block_style_arg(block_style),
             ),
             get_adaptive_pool(
                 option=adaptive_pool,
@@ -100,6 +107,24 @@ class SimpleConvNet2d(Sequential):
                 **self._dn_args,
             ),
         )
+
+    def get_features(self, batch: Tensor) -> Tensor:
+        """Get the features produced by the encoder and adaptive poool.
+
+        Parameters
+        ----------
+        batch : Tensor
+            A mini-batch of image-like inputs.
+
+        Returns
+        -------
+        Tensor
+            The features for each item in ``batch``.
+
+        """
+        feats = self[0](batch)
+        feats = self[1](feats)
+        return self[2](feats)
 
     _dn_args: Dict[str, Any] = {
         "hidden_sizes": None,