+++ tests/models/test_simple_conv_2d_call.py
@@ -16,6 +16,7 @@ def test_simple_conv_net_2d_call():
     adaptive_pool = ["avg", "max", "avg-max-concat"]
     lr_slopes = [0.0, 0.1]
     sizes = [1, 3, 5]
+    block_styles = ["double_conv", "conv_res"]
 
     iterator = product(
         ins,
@@ -26,9 +27,20 @@ def test_simple_conv_net_2d_call():
         adaptive_pool,
         lr_slopes,
         sizes,
+        block_styles,
     )
 
-    for in_chans, out_feats, feats, num_blocks, d_pool, a_pool, slope, size in iterator:
+    for (
+        in_chans,
+        out_feats,
+        feats,
+        num_blocks,
+        d_pool,
+        a_pool,
+        slope,
+        size,
+        block_style,
+    ) in iterator:
         model = SimpleConvNet2d(
             in_chans,
             out_feats,
@@ -38,6 +50,25 @@ def test_simple_conv_net_2d_call():
             adaptive_pool=a_pool,
             lr_slope=slope,
             kernel_size=size,
+            block_style=block_style,
         )
 
         assert model(rand(10, in_chans, 100, 100)).shape == (10, out_feats)
+
+
+def test_the_returned_shapes_of_get_features_method():
+    """Test the shapes returned by the ``get_features`` method."""
+    for feats_start, num_blocks in zip([16, 32], [3, 4, 5]):
+        model = SimpleConvNet2d(
+            3,
+            10,
+            features_start=feats_start,
+            num_blocks=num_blocks,
+        )
+
+        batch = rand(10, 3, 100, 100)
+
+        assert model.get_features(batch).shape == (
+            10,
+            feats_start * (2 ** (num_blocks - 1)),
+        )