+++ tests/models/test_variational_autoencoder_2d_contents.py
@@ -1,8 +1,10 @@
 """Test the contents of ``VAE2d``."""
-from torch.nn import Module
+from torch.nn import Module, Conv2d, LeakyReLU, BatchNorm2d
 
 from torch_tools import VAE2d
-from torch_tools.models._blocks_2d import DownBlock
+from torch_tools.models._blocks_2d import DownBlock, ConvBlock, ConvResBlock
+from torch_tools.models._blocks_2d import ResidualBlock, DoubleConvBlock
+from torch_tools.models._blocks_2d import UpBlock
 
 
 def test_number_of_blocks_in_encoder_and_decoder():
@@ -178,3 +180,193 @@ def test_vae_2d_contents_with_different_min_up_feats():
         )
 
         model.decoder.apply(lambda x: down_max_feats(x, min_feats))
+
+
+def test_vae_2d_encoder_contents_with_conv_res_blocks():
+    """Test the contents of the encoder in ``VAE2d`` with conv res blocks."""
+    model = VAE2d(
+        in_chans=3,
+        out_chans=3,
+        start_features=32,
+        input_dims=(64, 64),
+        block_style="conv_res",
+        lr_slope=0.123,
+    )
+
+    encoder = model.encoder
+
+    in_chans, out_chans = 32, 64
+
+    for block in list(encoder.children())[1:]:
+        assert isinstance(block, DownBlock)
+        assert isinstance(block[1], ConvResBlock)
+        assert isinstance(block[1][0], ConvBlock)
+
+        assert isinstance(block[1][0][0], Conv2d)
+        assert isinstance(block[1][0][1], BatchNorm2d)
+        assert isinstance(block[1][0][2], LeakyReLU)
+
+        assert block[1][0][0].in_channels == in_chans
+        assert block[1][0][0].out_channels == out_chans
+        assert block[1][0][1].num_features == out_chans
+        assert block[1][0][2].negative_slope == 0.123
+
+        assert isinstance(block[1][1], ResidualBlock)
+        assert isinstance(block[1][1].first_conv, ConvBlock)
+        assert isinstance(block[1][1].first_conv[0], Conv2d)
+        assert isinstance(block[1][1].first_conv[1], BatchNorm2d)
+        assert isinstance(block[1][1].first_conv[2], LeakyReLU)
+
+        assert block[1][1].first_conv[0].in_channels == out_chans
+        assert block[1][1].first_conv[0].out_channels == out_chans
+        assert block[1][1].first_conv[1].num_features == out_chans
+        assert block[1][1].first_conv[2].negative_slope == 0.0
+
+        assert isinstance(block[1][1].second_conv, ConvBlock)
+        assert isinstance(block[1][1].second_conv[0], Conv2d)
+        assert isinstance(block[1][1].second_conv[1], BatchNorm2d)
+
+        assert block[1][1].second_conv[0].in_channels == out_chans
+        assert block[1][1].second_conv[0].out_channels == out_chans
+        assert block[1][1].second_conv[1].num_features == out_chans
+
+        in_chans *= 2
+        out_chans *= 2
+
+
+def test_vae_2d_encoder_contents_with_double_conv_blocks():
+    """Test the contents of the encoder in ``VAE2d`` with conv res blocks."""
+    model = VAE2d(
+        in_chans=3,
+        out_chans=3,
+        start_features=32,
+        input_dims=(64, 64),
+        block_style="double_conv",
+        lr_slope=0.123,
+    )
+
+    encoder = model.encoder
+
+    in_chans, out_chans = 32, 64
+
+    for block in list(encoder.children())[1:]:
+        assert isinstance(block, DownBlock)
+        assert isinstance(block[1], DoubleConvBlock)
+        assert isinstance(block[1][0], ConvBlock)
+
+        assert isinstance(block[1][0][0], Conv2d)
+        assert isinstance(block[1][0][1], BatchNorm2d)
+        assert isinstance(block[1][0][2], LeakyReLU)
+
+        assert block[1][0][0].in_channels == in_chans
+        assert block[1][0][0].out_channels == out_chans
+        assert block[1][0][1].num_features == out_chans
+        assert block[1][0][2].negative_slope == 0.123
+
+        assert isinstance(block[1][1][0], Conv2d)
+        assert isinstance(block[1][1][1], BatchNorm2d)
+        assert isinstance(block[1][1][2], LeakyReLU)
+
+        assert block[1][1][0].in_channels == out_chans
+        assert block[1][1][0].out_channels == out_chans
+        assert block[1][1][1].num_features == out_chans
+        assert block[1][1][2].negative_slope == 0.123
+
+        in_chans *= 2
+        out_chans *= 2
+
+
+def test_vae_2d_decoder_contents_with_double_conv_blocks():
+    """Test the contents of the decoder in ``VAE2d`` with conv res blocks."""
+    model = VAE2d(
+        in_chans=3,
+        out_chans=3,
+        start_features=32,
+        input_dims=(64, 64),
+        block_style="double_conv",
+        lr_slope=0.123,
+        num_layers=3,
+    )
+
+    decoder = model.decoder
+
+    in_chans, out_chans = 32 * (2**2), 32 * (2**1)
+
+    for block in list(decoder.children())[:-1]:
+        assert isinstance(block, UpBlock)
+        assert isinstance(block[1], DoubleConvBlock)
+        assert isinstance(block[1][0], ConvBlock)
+
+        assert isinstance(block[1][0][0], Conv2d)
+        assert isinstance(block[1][0][1], BatchNorm2d)
+        assert isinstance(block[1][0][2], LeakyReLU)
+
+        assert block[1][0][0].in_channels == in_chans
+        assert block[1][0][0].out_channels == out_chans
+        assert block[1][0][1].num_features == out_chans
+        assert block[1][0][2].negative_slope == 0.123
+
+        assert isinstance(block[1][1][0], Conv2d)
+        assert isinstance(block[1][1][1], BatchNorm2d)
+        assert isinstance(block[1][1][2], LeakyReLU)
+
+        assert block[1][1][0].in_channels == out_chans
+        assert block[1][1][0].out_channels == out_chans
+        assert block[1][1][1].num_features == out_chans
+        assert block[1][1][2].negative_slope == 0.123
+
+        in_chans //= 2
+        out_chans //= 2
+
+
+def test_vae_2d_decoder_contents_with_conv_res_blocks():
+    """Test the contents of the decoder in ``VAE2d`` with conv res blocks."""
+    model = VAE2d(
+        in_chans=3,
+        out_chans=3,
+        start_features=32,
+        input_dims=(64, 64),
+        block_style="conv_res",
+        lr_slope=0.123,
+        num_layers=3,
+    )
+
+    decoder = model.decoder
+
+    in_chans, out_chans = 32 * (2**2), 32 * (2**1)
+
+    for block in list(decoder.children())[:-1]:
+        assert isinstance(block, UpBlock)
+        assert isinstance(block[1], ConvResBlock)
+        assert isinstance(block[1][0], ConvBlock)
+
+        assert isinstance(block[1][0][0], Conv2d)
+        assert isinstance(block[1][0][1], BatchNorm2d)
+        assert isinstance(block[1][0][2], LeakyReLU)
+
+        assert block[1][0][0].in_channels == in_chans
+        assert block[1][0][0].out_channels == out_chans
+        assert block[1][0][1].num_features == out_chans
+        assert block[1][0][2].negative_slope == 0.123
+
+        assert isinstance(block[1][1], ResidualBlock)
+        assert isinstance(block[1][1].first_conv, ConvBlock)
+        assert isinstance(block[1][1].first_conv[0], Conv2d)
+        assert isinstance(block[1][1].first_conv[1], BatchNorm2d)
+        assert isinstance(block[1][1].first_conv[2], LeakyReLU)
+
+        assert block[1][1].first_conv[0].in_channels == out_chans
+        assert block[1][1].first_conv[0].out_channels == out_chans
+        assert block[1][1].first_conv[1].num_features == out_chans
+        assert block[1][1].first_conv[2].negative_slope == 0.0
+
+        assert isinstance(block[1][1].second_conv, ConvBlock)
+        assert isinstance(block[1][1].second_conv[0], Conv2d)
+        assert isinstance(block[1][1].second_conv[1], BatchNorm2d)
+
+        assert block[1][1].second_conv[0].in_channels == out_chans
+        assert block[1][1].second_conv[0].out_channels == out_chans
+        assert block[1][1].second_conv[1].num_features == out_chans
+
+        in_chans //= 2
+        out_chans //= 2