+++ CHANGELOG.md
@@ -1,5 +1,15 @@
-## Version ?
+## Version 0.8.0
+  - Add ``get_features`` method to ``SimpleConvNet2d`` for extracting encoded features.
+  - Add ``get_features`` method to ``ConvNet2d`` for extracting encoded features.
   - Added a demo using a multiple-instance-learning attention model.
+
+## Version 0.7.0
+  - Changed ``AutoEncoder2d`` demo to use ovarian histology images.
+
+## Version 0.6.1
+  - Added residual blocks as optional block style to all relevant models.
+  - Changed the ``UNet`` demo to use a nuclei segmentation data set.
+>>>>>>> dev
 
 ## Version 0.6.0
   - Added synthetic shapes dataset and demo.